tree-wide: compat to current gyroidos wic-based build
Introduced necessary changes to be compatible with current meta-trustx
and used wic build. Further, we switch to current v6.12 branch of the
raspberrypi kernel.

The cml update fakeOS is not complete yet. It just contains the kernel
and CML but not the bootloader files and device tree files which are
located in the boot partition.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>