scripts/lib/wic/plugins/source/bootimg-partition-rpi-sb: Add
Add a custom wic plugin that is derived from and comaptible to
bootimg-partition. Instead of copying all files defined in
IMAGE_BOOT_FILES directly to the boot partition, all files are copied to
a 180MB FAT image which is subsequently signed for secure boot. Image
and signature are stored in the boot partition.

Signed-off-by: Johannes Wiesboeck <johannes.wiesboeck@aisec.fraunhofer.de>