images/gyroidos-cml-initramfs: Do not install openssh
This is a hack. The complete boot.img for RPi secure boot is limited to 180 MB.
We exceed this limit currently in dev builds. To comply with the limit,
we disable the CML SSH feature.

Perspectively, this issue should be solved by compressing the
kernel/initramfs bundle, using a intermediate bootloader/EFI, or
stripping down the kernel config.

Signed-off-by: Johannes Wiesboeck <johannes.wiesboeck@aisec.fraunhofer.de>